project/requirements.txt
pydealer-1.4.0
cowsay-5.0